+++ src/challenges/chess-simp/2024-05.ts
@@ -12,7 +12,9 @@ export class Simp_2024_05_01 implements Challenge {
     link: 'https://www.youtube.com/watch?v=ZM_8zFV9th8',
     challenge:
       "Chess, but it's a platformer. Your pieces (not pawns) must always land directly above another piece/pawn.",
-    records: new Map([[users.SenorFlamingo.name, { when: new Date('2024-05-06'), depth: 1 }]]),
+    records: new Map([
+      [users.SenorFlamingo.name, { when: new Date('2024-05-06'), depth: 1, moves: 17 }],
+    ]),
   }
 
   isMoveAllowed: Challenge['isMoveAllowed'] = ({ board, move }) => {